Extract URL resource construction into profile-logic.js so that it can be used by the chrome importer.
